chore: refactor feature product
